Altering User service and controller functions to work with JPA and the live database

+++ database/migrations/V1.0__initial_schema.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS public.user_profile
 
 CREATE TABLE IF NOT EXISTS public.users
 (
-    id serial constraint users_pk primary key,
+    id bigserial constraint users_pk primary key,
     user_name text not null UNIQUE,
     email_address text not null UNIQUE,
     user_password text not null,
@@ -16,20 +16,20 @@ CREATE TABLE IF NOT EXISTS public.users
 
 CREATE TABLE IF NOT EXISTS public.character_body_part_image_types
 (
-    id serial constraint character_body_part_image_types_pk primary key,
+    id bigserial constraint character_body_part_image_types_pk primary key,
     name text not null
 );
 
 CREATE TABLE IF NOT EXISTS public.character_body_part_images
 (
-    id serial constraint character_body_part_images_pk primary key,
+    id bigserial constraint character_body_part_images_pk primary key,
     url text not null,
     character_body_part_type int constraint character_body_part_types_fk references character_body_part_image_types(id)
 );
 
 CREATE TABLE IF NOT EXISTS public.character_images
 (
-    id serial constraint character_images_pk primary key,
+    id bigserial constraint character_images_pk primary key,
     head int constraint head_image_fk references character_body_part_images(id),
     body int constraint body_image_fk references character_body_part_images(id),
     legs int constraint legs_image_fk references character_body_part_images(id)
@@ -37,20 +37,20 @@ CREATE TABLE IF NOT EXISTS public.character_images
 
 CREATE TABLE IF NOT EXISTS public.background_images
 (
-    id serial constraint background_images_pk primary key,
+    id bigserial constraint background_images_pk primary key,
     image_name text not null,
     url text not null
 );
 
 CREATE TABLE IF NOT EXISTS public.prompts
 (
-    id serial constraint prompts_pk primary key,
+    id bigserial constraint prompts_pk primary key,
     adjectives text not null
 );
 
 CREATE TABLE IF NOT EXISTS public.prompts_collection
 (
-    id serial constraint prompts_collection_pk primary key,
+    id bigserial constraint prompts_collection_pk primary key,
     prompt1 text not null,
     prompt2 text not null,
     prompt3 text not null
@@ -58,7 +58,7 @@ CREATE TABLE IF NOT EXISTS public.prompts_collection
 
 CREATE TABLE IF NOT EXISTS public.characters
 (
-    id serial constraint characters_pk primary key,
+    id bigserial constraint characters_pk primary key,
     name text not null,
     prompts_collection_id int constraint prompts_collection_id_fk references public.prompts_collection(id),
     character_images_id int constraint character_images_id_fk references character_images(id),
@@ -67,12 +67,12 @@ CREATE TABLE IF NOT EXISTS public.characters
 
 CREATE TABLE IF NOT EXISTS public.catalogue
 (
-    id serial constraint catalogue_pk primary key,
+    id bigserial constraint catalogue_pk primary key,
     syllables text not null
 );
 
 CREATE TABLE IF NOT EXISTS public.character_catalogue
 (
-    id serial constraint character_catalogue_pk primary key,
+    id bigserial constraint character_catalogue_pk primary key,
     character_id int constraint character_id_fk references public.characters(id)
 );
